Removed non generic double
