+++ ArchiveViewer/Essence.Core/Commands/BaseCommand.cs
@@ -3,11 +3,11 @@ using System.Windows.Input;
 
 namespace Essence.Core.Commands
 {
-  public abstract class BaseCommand : ICommand
+  public abstract class BaseCommand<T> : ICommand
   {
-    bool ICommand.CanExecute(object parameter) => CanExecute();
+    bool ICommand.CanExecute(object parameter) => CanExecute(!(parameter is T obj) ? default (T) : obj);
 
-    void ICommand.Execute(object parameter) => Execute();
+    void ICommand.Execute(object parameter) => Execute(!(parameter is T obj) ? default (T) : obj);
 
     public event EventHandler CanExecuteChanged;
 
@@ -19,8 +19,8 @@ namespace Essence.Core.Commands
       canExecuteChanged((object) this, EventArgs.Empty);
     }
 
-    public abstract bool CanExecute();
+    public abstract bool CanExecute(T parameter);
 
-    public abstract void Execute();
+    public abstract void Execute(T parameter);
   }
 }
